backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
dataclasses-json==0.5.14
decorator==5.1.1
distlib==0.3.7
dnspython==2.4.2
entrypoints==0.4
filelock==3.12.2
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
frozenlist==1.3.3
gevent==23.7.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==21.0.1
idna==3.4
importlib-metadata==6.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
langchain==0.0.273
langsmith==0.0.26
llvmlite==0.40.1
loguru==0.7.2
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.20.1
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.57.1
numexpr==2.8.5
numpy==1.24.2
openai==0.27.4
openai-whisper @ git+https://github.com/openai/whisper.git@e8622f9afc4eba139bf796c210f5c01081000472
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pinecone-client==2.2.2
pipenv==2023.7.23
platformdirs==3.10.0
protobuf==3.20.3
pyarrow==11.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.15.0
Pympler==1.0.1
pyperclip==1.8.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.6.1
python-socketio==5.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.3.4
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
streamlit==1.21.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.3.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.1
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.4
validators==0.20.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
websocket==0.2.1
websocket-client==1.6.2
Werkzeug==2.3.7
yarl==1.8.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
